requirements.txt
discord.py[voice]
sqlalchemy